move to rules and add new hook
