+++ src/hooks/useIsInDragOverlay.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { DragOverlayRenderContext } from "@/components/DragOverlay/DragOverlay";
+
+const useIsInDragOverlay = () => useContext(DragOverlayRenderContext);
+
+export default useIsInDragOverlay;
